backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
prophet>=1.1.4
celery>=5.3.0
redis>=5.0.0
